+++ src/data/wars/quadring/war.ts
@@ -1,0 +1,8 @@
+import type { WarCreate } from "../../War";
+
+export const data: WarCreate = {
+  name: "Quadring",
+  number: 4,
+  startTime: new Date("10 feb 2023"),
+  endTime: new Date("26 mars 2023"),
+};
