+++ src/data/wars/we-go-again/problems/human-tings/data.ts
@@ -1,16 +1,15 @@
 import type { Problem } from "../../../../Problem";
-  
+
 export const data: Problem<string, string> = {
-    name: "Human tings",
-    difficulty: "easy",
-    number: 63,
-    arguments: [{name: "s", type: "string"}],
-    tags: ["basics", "numbers", "objects"],
-    testCases: [
-      {
-        input: [`"s"`],
-        output: "a",
-      },
-    ],
-  };
-    
+  name: "Human tings",
+  difficulty: "easy",
+  number: 63,
+  arguments: [{ name: "s", type: "string" }],
+  tags: ["basics", "numbers", "objects"],
+  testCases: [
+    {
+      input: [`"s"`],
+      output: "a",
+    },
+  ],
+};
